+++ flixopt/__init__.py
@@ -17,7 +17,7 @@ except (PackageNotFoundError, TypeError):
 # - xr.Dataset.fxstats (from stats_accessor)
 import xarray_plotly as _xpx  # noqa: F401
 
-from . import clustering, linear_converters, plotting, results, solvers, tutorials
+from . import clustering, linear_converters, plotting, solvers, tutorials
 from . import stats_accessor as _fxstats  # noqa: F401
 from .carrier import Carrier, CarrierContainer
 from .comparison import Comparison
@@ -36,7 +36,6 @@ from .elements import Bus, Flow
 from .flow_system import FlowSystem
 from .flow_system_status import FlowSystemStatus
 from .interface import InvestParameters, Piece, Piecewise, PiecewiseConversion, PiecewiseEffects, StatusParameters
-from .optimization import Optimization, SegmentedOptimization
 from .plot_result import PlotResult
 
 __all__ = [
@@ -57,8 +56,6 @@ __all__ = [
     'Transmission',
     'FlowSystem',
     'FlowSystemStatus',
-    'Optimization',
-    'SegmentedOptimization',
     'InvestParameters',
     'StatusParameters',
     'Piece',
@@ -68,7 +65,6 @@ __all__ = [
     'PlotResult',
     'clustering',
     'plotting',
-    'results',
     'linear_converters',
     'solvers',
     'tutorials',